@@ -3,41 +3,19 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@dustinirving/component-library": "^1.0.2",
+    "@mdi/js": "^7.1.96",
+    "@mdi/react": "^1.6.1",
     "@mdi/svg": "^6.9.96",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "4.0.3",
-    "styled-components": "^5.3.5"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "styled-components": "^5.3.5",
+    "tailwind-component-library": "file:../tailwind-component-library/tailwind-component-library-1.0.0.tgz"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "prepare": "husky install",
-    "pretty-quick": "pretty-quick"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview",
+    "start": "npm run dev"
   },
   "devDependencies": {
     "@types/jest": "^28.1.4",
@@ -45,9 +23,15 @@
     "@types/react": "^18.0.14",
     "@types/react-dom": "^18.0.5",
     "@types/styled-components": "^5.1.25",
+    "@vitejs/plugin-react": "^3.0.0",
+    "autoprefixer": "^10.4.13",
     "husky": "^7.0.2",
+    "postcss": "^8.4.20",
     "prettier": "^2.3.2",
     "pretty-quick": "^3.1.1",
-    "typescript": "^4.7.4"
+    "tailwindcss": "^3.2.4",
+    "typescript": "^4.7.4",
+    "vite": "^4.0.3",
+    "vite-plugin-svgr": "^2.4.0"
   }
 }
